ДЗ 7. перевод репозиториев на spring-data-jpa JpaRepository
