--- homework06/src/main/java/com/example/homework06/dto/BookDto.java
@@ -1,13 +1,0 @@
-package com.example.homework06.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class BookDto {
-    private Long id;
-    private String name;
-    private Long authorId;
-    private Long genreId;
-}
