--- homework06/src/main/java/com/example/homework06/dto/CommentDto.java
@@ -1,12 +1,0 @@
-package com.example.homework06.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class CommentDto {
-    private Long id;
-    private String text;
-    private Long bookId;
-}
