--- homework06/src/main/java/com/example/homework06/dto/CommentResponseDto.java
@@ -1,13 +1,0 @@
-package com.example.homework06.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class CommentResponseDto {
-    private Long id;
-    private String text;
-    private String bookName;
-    private String authorName;
-}
