--- homework06/src/main/java/com/example/homework06/exceptions/NotFoundException.java
@@ -1,7 +1,0 @@
-package com.example.homework06.exceptions;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
