--- homework06/src/main/java/com/example/homework06/model/Book.java
@@ -1,44 +1,0 @@
-package com.example.homework06.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.SequenceGenerator;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "BOOKS")
-public class Book {
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "books_generator")
-    @SequenceGenerator(name = "books_generator",sequenceName = "BOOKS_SEQ",  allocationSize = 1)
-    @Column(name = "BOOK_ID")
-    private Long id;
-
-    @Column(name = "NAME")
-    private String name;
-
-    @ManyToOne
-    @JoinColumn(name = "AUTHOR_ID", referencedColumnName = "AUTHOR_ID")
-    private Author author;
-
-    @ManyToOne
-    @JoinColumn(name = "GENRE_ID", referencedColumnName = "GENRE_ID")
-    private Genre genre;
-
-    public Book(String name, Author author, Genre genre) {
-        this.name = name;
-        this.author = author;
-        this.genre = genre;
-    }
-}
