+++ homework07/src/main/java/com/example/homework07/Main.java
@@ -1,0 +1,16 @@
+package com.example.homework07;
+
+
+import java.sql.SQLException;
+import org.h2.tools.Console;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ApplicationContext;
+
+@SpringBootApplication
+public class Main {
+    public static void main(String[] args) throws SQLException {
+        ApplicationContext context = SpringApplication.run(Main.class, args);
+        Console.main(args);
+    }
+}
