+++ homework07/src/main/java/com/example/homework07/dto/BookDto.java
@@ -1,0 +1,13 @@
+package com.example.homework07.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class BookDto {
+    private Long id;
+    private String name;
+    private Long authorId;
+    private Long genreId;
+}
