+++ homework07/src/main/java/com/example/homework07/dto/CommentDto.java
@@ -1,0 +1,12 @@
+package com.example.homework07.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class CommentDto {
+    private Long id;
+    private String text;
+    private Long bookId;
+}
