+++ homework07/src/main/java/com/example/homework07/exceptions/NotFoundException.java
@@ -1,0 +1,7 @@
+package com.example.homework07.exceptions;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
