+++ homework07/src/main/java/com/example/homework07/repository/AuthorRepository.java
@@ -1,0 +1,8 @@
+package com.example.homework07.repository;
+
+import com.example.homework07.model.Author;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AuthorRepository extends JpaRepository<Author, Long> {
+}
