+++ homework07/src/main/java/com/example/homework07/repository/BookRepository.java
@@ -1,0 +1,7 @@
+package com.example.homework07.repository;
+
+import com.example.homework07.model.Book;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookRepository extends JpaRepository<Book, Long> {
+}
