+++ homework07/src/main/java/com/example/homework07/repository/CommentRepository.java
@@ -1,0 +1,10 @@
+package com.example.homework07.repository;
+
+import com.example.homework07.model.Comment;
+import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+
+    List<Comment> findAllByBookId(long id);
+}
