+++ homework07/src/main/java/com/example/homework07/repository/GenreRepository.java
@@ -1,0 +1,7 @@
+package com.example.homework07.repository;
+
+import com.example.homework07.model.Genre;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface GenreRepository extends JpaRepository<Genre, Long> {
+}
