+++ homework07/src/main/resources/schema.sql
@@ -1,0 +1,28 @@
+CREATE TABLE IF NOT EXISTS AUTHORS
+(
+    AUTHOR_ID BIGINT PRIMARY KEY,
+    name      VARCHAR(255)
+);
+CREATE TABLE IF NOT EXISTS GENRES
+(
+    GENRE_ID BIGINT PRIMARY KEY,
+    name     VARCHAR(255)
+);
+CREATE TABLE IF NOT EXISTS BOOKS
+(
+    BOOK_ID   BIGINT PRIMARY KEY,
+    name      VARCHAR(255),
+    AUTHOR_ID BIGINT,
+    GENRE_ID  BIGINT,
+    FOREIGN KEY (AUTHOR_ID) REFERENCES AUTHORS (AUTHOR_ID),
+    FOREIGN KEY (GENRE_ID) REFERENCES GENRES (GENRE_ID)
+);
+CREATE TABLE IF NOT EXISTS COMMENTS
+(
+    COMMENT_ID BIGINT PRIMARY KEY,
+    text       VARCHAR(255),
+    BOOK_ID    BIGINT,
+    FOREIGN KEY (BOOK_ID) REFERENCES BOOKS (BOOK_ID)
+);
+CREATE SEQUENCE IF NOT EXISTS "BOOKS_SEQ" START WITH 3;
+CREATE SEQUENCE IF NOT EXISTS "COMMENTS_SEQ" START WITH 3
